refactor: move CLI commands/parsing into one module
